HOTFIX: Se corrije tipo de salida de metodo de Manager y se corrije makefile

+++ Src/BlockChainManager.cpp
@@ -107,6 +107,6 @@ void BlockChainManager::setUserDefinedDifficulty(int d){
 }
 
 
-size_t BlockChainManager::getUserDefinedDifficulty( void ){
+unsigned int BlockChainManager::getUserDefinedDifficulty( void ){
 	return userDefinedDifficulty;
 }
